+++ rf4s/config/defaults.py
@@ -104,6 +104,9 @@ _C.BOT.CLICK_LOCK = False
 _C.BOT.KEEPNET = CN()
 # Capacity of the keepnet
 _C.BOT.KEEPNET.CAPACITY = 100
+# Events that triggers a screenshot.
+# (options: fish, card, gift)
+_C.BOT.KEEPNET.SCREENSHOT_EVENT = ("fish", "card", "gift")
 # Take a screenshot when capturing a fish with any tags below.
 # Always take a screenshot if it's left empty.
 # (options: green, yellow, blue, purple, pink)