+++ docs/en/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix pause bug
 - Optimize time required to pause
 - Now reset fb on quitting
+- Change default pause fb reset and quit key
 
 
 ## 0.6.3 (2025-08-11)