+++ docs/en/CHANGELOG.md
@@ -11,7 +11,7 @@
 - Change default pause fb reset and quit key
 - Add new marine profiles
 - Rename old profiles
-- Add SCREENSHOT_EVENT
+- Add SCREENSHOT_EVENTS
 
 
 ## 0.6.3 (2025-08-11)